Updated for heroku.

+++ repository/src/main/java/ru/mail/agb88/repository/model/Sensor.java
@@ -11,24 +11,24 @@ public class Sensor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column (name = "F_ID", insertable = false, updatable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "F_VALUE")
-    private Integer value;
+    private Long value;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getValue() {
+    public Long getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Long value) {
         this.value = value;
     }
 
@@ -36,8 +36,8 @@ public class Sensor {
     }
 
     public static final class Builder {
-        private Integer id;
-        private Integer value;
+        private Long id;
+        private Long value;
 
         private Builder() {
         }
@@ -46,12 +46,12 @@ public class Sensor {
             return new Builder();
         }
 
-        public Builder id(Integer id) {
+        public Builder id(Long id) {
             this.id = id;
             return this;
         }
 
-        public Builder value(Integer value) {
+        public Builder value(Long value) {
             this.value = value;
             return this;
         }
