+++ service/src/main/java/ru/mail/agb88/service/DTO/SensorDTO.java
@@ -4,28 +4,28 @@ package ru.mail.agb88.service.DTO;
  * Created by AlexBal 25.11.2017
  */
 public class SensorDTO {
-    private Integer id;
-    private Integer value;
+    private Long id;
+    private Long value;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getValue() {
+    public Long getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Long value) {
         this.value = value;
     }
 
     public static final class Builder {
-        private Integer id;
-        private Integer value;
+        private Long id;
+        private Long value;
 
         private Builder() {
         }
@@ -34,12 +34,12 @@ public class SensorDTO {
             return new Builder();
         }
 
-        public Builder id(Integer id) {
+        public Builder id(Long id) {
             this.id = id;
             return this;
         }
 
-        public Builder value(Integer value) {
+        public Builder value(Long value) {
             this.value = value;
             return this;
         }
